Work around iOS memory protection restrictions
iOS usually doesn't allow for JIT compilation, since it generally enforces W^X,
but this restriction can be lifted if you have get-task-allow. Even then, it
make sure that you don't try to have *both* write and execute privileges, so a
bit of a workaround is required where we set the execute bit right after
relocation and unset it right before deallocating.